Feat : 기업 입사제안시 실시간 알림

+++ src/main/java/com/sist/jobgem/controller/AlertController.java
@@ -19,12 +19,12 @@ public class AlertController {
     private final AlertService alertService;
 
     @GetMapping(value = "/subscribe/{userId}", produces = MediaType.TEXT_EVENT_STREAM_VALUE)
-    public SseEmitter subscribe(@PathVariable Long userId) {
+    public SseEmitter subscribe(@PathVariable int userId) {
         return sseService.subscribe(userId);
     }
 
     @GetMapping(value = "/send/{userId}/{data}")
-    public void send(@PathVariable Long userId, @PathVariable Object data) {
+    public void send(@PathVariable int userId, @PathVariable Object data) {
         sseService.sendToClient(userId, data);
     }
